# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AssignOrderRequest": {
        "properties": {
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "AssignOrderRequest",
        "type": "object"
      },
...
      "BlockedOn": {
        "description": "Which check a BLOCKED or NEEDS_OVERRIDE state is blocked on.",
        "enum": [
          "LOCATION",
          "SKU",
          "QTY"
        ],
        "title": "BlockedOn",
        "type": "string"
      },
      "ClipDescriptor": {
        "description": "One clip a device needs, and where to fetch its audio.",
        "properties": {
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "content_hash",
          "url",
          "text"
        ],
        "title": "ClipDescriptor",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NextLineResponse": {
        "properties": {
          "aisle": {
            "title": "Aisle",
            "type": "string"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "bay": {
            "title": "Bay",
            "type": "string"
          },
          "blocked_on": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlockedOn"
              },
              {
                "type": "null"
              }
            ]
          },
          "effects": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Effects",
            "type": "array"
          },
          "expected_qty": {
            "title": "Expected Qty",
            "type": "integer"
          },
          "instruction_text": {
            "title": "Instruction Text",
            "type": "string"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "line_id": {
            "format": "uuid",
            "title": "Line Id",
            "type": "string"
          },
          "location_barcode": {
            "title": "Location Barcode",
            "type": "string"
          },
          "order_id": {
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "sku_code": {
            "title": "Sku Code",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/PickState"
          }
        },
        "required": [
          "line_id",
          "order_id",
          "sequence",
          "state",
          "attempts",
          "blocked_on",
          "aisle",
          "bay",
          "level",
          "location_barcode",
          "sku_code",
          "reference",
          "expected_qty",
          "instruction_text",
          "effects"
        ],
        "title": "NextLineResponse",
        "type": "object"
      },
...
      "OrderResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "order_code": {
            "title": "Order Code",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          }
        },
        "required": [
          "id",
          "order_code",
          "status",
          "session_id"
        ],
        "title": "OrderResponse",
        "type": "object"
      },
      "OrderStatus": {
        "description": "Lifecycle of a pick order.",
        "enum": [
          "PENDING",
          "IN_PROGRESS",
          "COMPLETED",
          "CANCELLED"
        ],
        "title": "OrderStatus",
        "type": "string"
      },
      "PickState": {
        "description": "States a single pick line can be in.",
        "enum": [
          "PENDING",
          "AWAITING_LOCATION",
          "AWAITING_SCAN",
          "BLOCKED",
          "NEEDS_OVERRIDE",
          "AWAITING_QTY",
          "SHORT_PENDING",
          "DONE",
          "EXCEPTED"
        ],
        "title": "PickState",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VoiceManifestResponse": {
        "description": "Every clip a device needs prefetched for one order.",
        "properties": {
          "clips": {
            "items": {
              "$ref": "#/components/schemas/ClipDescriptor"
            },
            "title": "Clips",
            "type": "array"
          },
          "order_id": {
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          }
        },
        "required": [
          "order_id",
          "clips"
        ],
        "title": "VoiceManifestResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "macenplast-voice-picking",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/orders/{order_id}/assign": {
      "post": {
        "operationId": "assign_order_orders__order_id__assign_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Assign Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{order_id}/next-line": {
      "get": {
        "operationId": "get_next_line_orders__order_id__next_line_get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextLineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Next Line",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/voice/orders/{order_id}/manifest": {
      "get": {
        "description": "List every clip a device needs prefetched to run this order in VOICE\nmode, synthesizing (and caching) anything not already built.",
        "operationId": "get_voice_manifest_voice_orders__order_id__manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceManifestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Voice Manifest",
        "tags": [
          "voice"
        ]
      }
    },
...
  }
}